Add VLC Media Player to the templates
Hello! I've added VLC Media Player to the templates. I've created the new "MediaPlayer" class for it, and I'm planning on expanding that class with MPC-HC if you're OK with this pull request :)

I've also added an option for VLC to access the AppData in order to save the settings, and another one for access to the pictures folder to save screenshots.