Disable Turkish for inactivity
There is a concrete risk of a translation mismatch because of the continuos Plus UI changes. Since the latest Turkish update, more than 20 versions were released, so the Plus support for Turkish should be turned off.

If you want to update and restore Plus translation, please open a new issue.