Restore Waterfox phishing template
Removing the lines was not correct (#1309). The file "cert9.db" is still used in the current version of Waterfox. The file "blocklist.xml" is probably still used by older versions.

The problem was not because these files supposedly no longer exist, but because the variable could not be expanded due to a typo (capitalized F).