Add helper for setting tab orders automatically
The helper functions are put into separate files under the Helpers folder.

The helper is used in OptionsWindow and SettingsWindow.

The helper sets the tab order of a widget using its row and column number in the grid layout. The tab orders are assigned first from left to right, then from top to bottom.

If in the future a different custom type of layout is used, or a container other than QTabWidget is used, then the helper functions may need to be changed.