Add configurable tooltips to code completion popup
Introduces a tooltip callback mechanism and a custom item delegate to show tooltips for completion items in the code editor popup. Tooltips are now configurable via settings and can be enabled or disabled globally. Updates OptionsWindow and SettingsWindow to use the new tooltip callback and respect the popup tooltip setting.